requirements.txt
torch==1.11.0
numpy==1.19.2
pyyaml==5.4.1
Flask==2.1.2
nltk==3.5
pandas==1.2.3
faiss-cpu==1.7.0
scikit_learn==1.0
langdetect==1.0.9
requests==2.28.1
ipython==8.4.0
